// src/types/schemas-auth.d.ts
export interface components {
    schemas: {
// ...
        RefreshTokenRequest: {
            refreshToken: string;
        };
        ApiResponseRefreshTokenResponse: {
            success?: boolean;
            data?: components["schemas"]["RefreshTokenResponse"];
            message?: string;
            errorCode?: string;
            /** Format: date-time */
            timestamp?: string;
        };
        RefreshTokenResponse: {
            accessToken?: string;
            refreshToken?: string;
            tokenType?: string;
            /** Format: int64 */
            accessExpiresInSeconds?: number;
            /** Format: int64 */
            refreshExpiresInSeconds?: number;
            /** Format: date-time */
            refreshExpiresAt?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}